Update patch set 1
Patch Set 1:

> The other reason this is useful is that if for any other reason you have parsed the Leaf before we should use it.
There is a potential improvement in X509KeyPair() for example, where the leaf cert is parsed but not preserved. See https://golang.org/src/crypto/tls/tls.go?s=5915:5986#L242

> I don't see how this would be useful if you are using GetCertificate, which ignores Certificates, did you mean GetConfigForClient? Note though that GetCertificate does let you maintain an out-of-box name map, so you might want to switch to that callback.

Good to know. Functional-wise they are similar for my usage. GetCertificate() was chose at the beginning as it is straightforward.

Patch-set: 1