Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> Why not by default have exactly the same behavior as we have now.
> 
> I share Russ' concern that this may result in unnecessary failures. Doing anything during normal operation assumes that invariants are themselves not bug free.

In an in-person discussion, Russ suggested that if we're actually sure these invariants hold, we should just put in panics. He's concerned that avoiding them would encourage users to add invariant checks as a way of testing things they're not sure about, rather than simplifying the code so that they can be confident one way or the other (and then put in a panic that also applies in the production path).

I'm not so sure I buy that, but for now I think I'm going to add panics to the `go` command and see how far we get.


> But it makes sense to get the extra information during testing.
> 
> Would it not make sense to test the new API internally first and see how it works out, rather than adding the invariant API right off the bat?

Indeed, that's why I put it in cmd/go/internal. 🙂

Patch-set: 5
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>