Update patch set 3
Patch Set 3:

(4 comments)

This is a nice idea, thanks. In addition to the comments I made (take 'em or leave 'em), I have a mild concern that we'll want to use this functionality to start network servers, and this won't be that useful in that case, because either we choose the network port in advance (leaving the risk that some other process grabs it before the process has started listening on it) or we let the process itself choose it, in which case there's no way to find out what the port is, because there's no way to process stdout or stderr before the process has finished, even if we did have a command that could do that.

Patch-set: 3