runtime: eagerly scavenge based on global budget
Previously, we would always scavenge on heap growth by however much
sysAlloc would map. However, this is perhaps too aggressive, and results
in scavenging more than we need every time we grow the heap. This can
result in long-term TLB misses.

Instead, in this change, we manage a global budget. Each time we
use any scavenged space (note that a span created from freshly grown
heap space is considered scavenged) we add that as scavenge debt. If
that debt exceeds scavengeDebtMax, then we eagerly scavenge the heap.

For #16930.

Change-Id: Ibdec4acd3e66df355f1a3b7c36e772a114671691