Update patch set 1
Patch Set 1:

Jan, please try to be more polite, especially to new contributors.  Thanks.

That said, I tend to agree that it is not productive to send small tweaks like this to existing code.  Occasionally it can make the code clearer.  In general, though it doesn't matter.  And it makes no difference to code performance.  We generally try to avoid making minor tweaks to existing working code.  If the code works, the only real effect that minor tweaks can have is break it.  Better to leave working code alone.

Thanks for your interest in this kind of work, but let's focus on cases where the code change really does make it more readable.  Thanks.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>