crypto/tls: replace custom *block with standard buffers
The crypto/tls record layer used a custom buffer implementation with its
own semantics, freelist, and offset management. Replace it all with
per-task bytes.Buffer, bytes.Reader and byte slices, along with a
refactor of all the encrypt and decrypt code.

The main quirk of *block was to do a best-effort read past the record
boundary, so that if a closeNotify was waiting it would be peeked and
surfaced along with the last Read. Address that with atLeastReader and
ReadFrom to avoid a useless copy (instead of a LimitReader and CopyN).

There was also an optimization to split blocks along record boundary
lines without having to copy in and out the data. Replicate that by
splitting a bytes.Buffer into a bytes.Reader for the head and a new
bytes.Buffer for the tail. The tail will get to use the spare capacity
and then will drop the underlying array when it needs to reallocate.
Interesting that "buf = *bytes.NewBuffer(slice)" doesn't allocate.

DO NOT SUBMIT pending benchmarks. There might be a slowdown due to the
c.input buffer getting dropped on the floor instead of reused.

This reduced the starting buffer from 1024 to 5 + 512, so it triggered

Change-Id: Ie31b0a572b2ad37878469877798d5c6a5276f931