Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Seems reasonable to me but since I'm completely unfamiliar with the schedule pass I'll let Keith do the actual review.
> 
> Are OpArgs always in b0? That'd be nice.

OpArgs are always in the entry block, but within that block (which can be largish) they were floated down closer to their uses.

Still TBD is whether delve is not honoring the "this is the prologue" information or if that dropped through the cracks in 1.11.  I also have a CL coming that sorts the variables so that the parameters come first, since that is what clang seems to do (but that doesn't make gdb work, either).

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>