Update patch set 1
Patch Set 1:

> Still TBD is whether delve is not honoring the "this is the prologue" information or if that dropped through the cracks in 1.11

Delve is honoring the prologue thing, it has been since before 1.11 by detecting the prologue in the disassembly.

If you are seeing problems on this front it's a bug in delve.

Regarding this CL if you look at fmt.Fprintln's DIE you see:

 <1><7353c>: Abbrev Number: 3 (DW_TAG_subprogram)
    <7353d>   DW_AT_name        : fmt.Fprintln
    <7354a>   DW_AT_low_pc      : 0x4858a0
    <73552>   DW_AT_high_pc     : 0x485992
    <7355a>   DW_AT_frame_base  : 1 byte block: 9c      (DW_OP_call_frame_cfa)
    <7355c>   DW_AT_decl_file   : 0x2
    <73560>   DW_AT_external    : 1
 <2><73561>: Abbrev Number: 11 (DW_TAG_variable)
    <73562>   DW_AT_name        : p
    <73564>   DW_AT_decl_line   : 263
    <73566>   DW_AT_type        : <0x40833>
    <7356a>   DW_AT_location    : 0x63fd3 (location list)
 <2><7356e>: Abbrev Number: 15 (DW_TAG_formal_parameter)
    <7356f>   DW_AT_name        : w
    <73571>   DW_AT_variable_parameter: 0
    <73572>   DW_AT_decl_line   : 262
    <73574>   DW_AT_type        : <0x408f8>
    <73578>   DW_AT_location    : 0 byte block:         ()
(...)

note the `DW_AT_location` of `w`, that's why delve is not displaying the value of the `w` parameter on entry.

Patch-set: 1
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>