Update patch set 1
Patch Set 1:

> Delve is honoring the prologue thing, it has been since before 1.11 by detecting the prologue in the disassembly.
> 
> If you are seeing problems on this front it's a bug in delve.

I think I do see a delve problem.  Nowadays, on amd64, we also construct a base pointer, so there are two more instructions to consider.  It would also "be nice" if stop-in-foo could be done by setting the breakpoint on the last instruction of foo's (modified) prologue, then step one more forward so it is "as if" the breakpoint were at the true first instruction. Because the true first instruction might be part of a loop, I'm not sure you want to put the actual breakpoint right there since it would fire on every loop iteration.
```
	file.go:332 0x10ac010 65488b0c2530000000 mov rcx, qword ptr gs:[0x30]
	file.go:332 0x10ac019 483b6110		 cmp rsp, qword ptr [rcx+0x10]
	file.go:332 0x10ac01d 0f86a4000000	 jbe 0x10ac0c7
=>	file.go:332 0x10ac023 4883ec38		 sub rsp, 0x38
	file.go:332 0x10ac027 48896c2430	 mov qword ptr [rsp+0x30], rbp
	file.go:332 0x10ac02c 488d6c2430	 lea rbp, ptr [rsp+0x30]
--------------------------
	file.go:333 0x10ac031 488b442440	 mov rax, qword ptr [rsp+0x40]
	file.go:333 0x10ac036 0f1000		 movups xmm0, xmmword ptr [rax]
```
The compiler/assembler/linker still conspire to begin the location list lifetime at  0x10ac036, which is one instruction too late; that is next on my list.  This all still looks broken with gdb.

```
0x0008890c: TAG_formal_parameter [16]  
             AT_name( "t" )
             AT_variable_parameter( 0x00 )
             AT_decl_line( 332 )
             AT_type( {0x0000000000042aab} ( *_/.../cmd/splitdwarf/macho.FileTOC ) )
             AT_location( 0x00082ad8
               0x00000000010ac036 - 0x00000000010ac0d1: call-frame-cfa )
```

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>