Update patch set 1
Patch Set 1:

> I think I do see a delve problem.  Nowadays, on amd64, we also construct a base pointer, so there are two more instructions to consider.  It would also "be nice" if stop-in-foo could be done by setting the breakpoint on the last instruction of foo's (modified) prologue, then step one more forward so it is "as if" the breakpoint were at the true first instruction. 

I don't understand what is the point of doing this. The reason we have for skipping the prologue is to avoid the loop that calls runtime.morestack_noctxt, nothing else. It makes no difference whether we stop on the first instruction past the jump or the last instruction of the prologue. 

Function arguments in principle should have a location starting at the entry point of the function.

Patch-set: 1
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>