Update patch set 1
Patch Set 1:

> In principle, yes they have that address, in practice it requires additional assembler/linker hackery because the compiler doesn't "see" any of that when it generates DWARF and location lists.  My goal is that if someone says "break foo", then when the debugger stops on entry to foo, all the parameters can be printed.  

I suspect those two goals (no additional assembler/linker hackery, printing the parameters at "foo" breakpoint) may be incompatible. 

We (delve) need the prologue to end on the same line as the "func" line. Our most used frontends are all IDE integrations, that means that people usually click on a source line to set a breakpoint. It would be weird if, when people clicked on the "func" line, the breakpoint appeared in the following line.

Besides user expectations, I'm not at all confident that if we started doing this it wouldn't break our frontends, which is something we try not to do.

> I'm also a little unclear on whether debuggers use the virtual frame pointer, or the actual frame pointer -- if we need to describe parameters at those points in execution and it is actual FP, then we'll need different entries, meaning more entries (meaning slightly larger executable).

In theory: neither. Stack variables are all represented as offset from the "CFA" (Call Frame Address). This is a fake DWARF register that is calculated using debug_frame in an arbitrarily complicated way. 

In practice CFA is the value of SP before the CALL instruction that created the current frame was executed. The linker already records the offset of CFA from SP in debug_frame (spdelta, basically). In practice, I think, CFA is the same as the virtual frame pointer (but I'm not 100% sure about what the virtual frame pointer is).

> 
> E.g. in
> 0x0008891a: TAG_formal_parameter [16]  
>              AT_name( "Type" )
>              AT_variable_parameter( 0x00 )
>              AT_decl_line( 332 )
>              AT_type( {0x0000000000042871} ( _/.../splitdwarf/macho.HdrType ) )
>              AT_location( 0x00082b0b
>                0x00000000010ac036 - 0x00000000010ac0d1: fbreg +8 )
> at those instructions before BP/FP is initialized, there is no fbreg.

fbreg always exists for go functions, it's defined in debug_info (with DW_AT_frame_base) as being the same as the CFA.

> 
> But maybe I am doing this the wrong way -- gdb persists in claiming that all the variables are optimized away.

That's weird. It should display them as long as you are inside the PC range.

Patch-set: 1
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>