flag: add description of default values to docs
During a recent go code school I helped with the concept of default
values with flags was a little confusing to some of the students.

Due to the fact they obtain a pointer value from flag.Parse() some
thought it would be possible to have an optional flag and just check
for nil.

I thought it would be useful to describe the fact the default value
is used.

Change-Id: Ibffbe8e13a3f88409e680cd439b080aa7e167cbc