Update patch set 38
Patch Set 38:

> Patch Set 38: -Code-Review
> 
> > 1 of 19 TryBots failed:
> > Failed on nacl-386: https://storage.googleapis.com/go-build-log/2bd42ae8/nacl-386_4b2b7f6c.log
> 
> This one looks real.

Yeah, unfortunately. This makes me sweaty, as it means that there's still a hole in treap access somewhere. I'll apply the same approach to the stack up to here as I am for the background scavenger. It's possible it's also the exact same issue, just that the background scavenger might be running in a wider variety of contexts.

Patch-set: 38
Reviewer: Gerrit User 25391 <25391@62eb7196-b449-3ce5-99f1-c037f21e1705>