cmd/go/internal/work: don't apply O_EXCL when probing for umask
It should be fine for two 'go build' commands to concurrently write
the same content to the same file. However, they use a temporary file
to probe for the umask of the directory, and could theoretically fail
to create that file if another go command created it concurrently (or,
worse, crashed after creating it but before deleting it).

Updates #26794

Change-Id: I5226c6e5d2b4f464a62bdb3460a635eac776ef0f