cmd/compile: force OpArg location list starts to beginning
For the entry block, make the "first instruction" be truly
the first instruction.  This allows printing of incoming
parameters with Delve.

Also be sure Phis are marked as being at the start of their
block.  This is observed to move location list pointers,
and where moved, there are more correct (need to be one
instruction earlier, either a correction to this CL
or a follow-on CL).

Change-Id: Ic043b7265835cf1790382a74334b5714ae4060af