Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> Do we want keep displaying successors indexes for edges? I cannot see the usefulness of it.

You are correct (in an early comment) that there is an issue with how backwards edges are presented.  An easy way to classify backedges is to use

ssa.postorderWithNumbering(f *Func, ponums []int32) []*Block

(you might need to make it public if you need to call it from gc/ssa.go)

as in

ponums := make([]int32, f.NumBlocks())
_ = postorderWithNumbering(f, ponums)

isbackedge := func(bfrom, bto *Block) boolean {
    return ponums[bfrom.ID] <= ponums[bto.ID]
}

"postorder" means that a block appears in the sequence before any of its (non-loop) predecessors; that is, forward flow goes from a higher numbered block (a predecessor, appearing later in the postorder) to a lower numbered block. A lower-or-equal-numbered predecessor must be a loop edge (and I am embarrassed at the dearth of explanatory comments around this function in the compiler source code).

Patch-set: 9
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>