crypto/tls: implement full Lucky 13 countermeasures for SHA-1 CBC ciphers
The ConstantTimeSum operation does not provide a full Lucky 13
countermeasure.

In TLS CBC padding, we have some unauthenticated input where some secret
prefix is the plaintext. We must compute HMAC on that prefix without
leaking its length.

ConstantTimeSum attempts to protect d.nx, or plaintextLen % BlockSize.
This value is indeed secret, but it was set by a preceding
Write(in[:plaintextLen]) call which took no precautions.

While we could implement a ConstantTimeWrite, it would need to treat
d.nx secret to account for the caller calling ConstantTimeWrite several
times in a row. This is complexity which is not needed by TLS. In TLS,
the only secret length is the final Write() call.

Instead, we replace ConstantTimeSum with ConstantTimeSumWithData, which
incorporates a final secret-length Write into the hash but assumes all
prior Write lengths are public. This is sufficient to implement the
mitigations described in
https://www.imperialviolet.org/2013/02/04/luckythirteen.html

This change leaves the SHA-256 ciphers with an analog of the previous
implementation. SHA-256 ciphers are disabled by default and require TLS
1.2, at which point the modern AEAD-based ciphers are available.

CBC-mode ciphers continue to be considered obsolete. Use AEAD-based
ciphers instead.

Fixes #27071

Change-Id: I3f2b9ffe25e43f15dc441afb752a03ea8118c49a