text/template/parse: skip leading BOM character
Some editors insert this character at the beginning of files, which is
already skipped in go/scanner and cmd/compile when reading Go source
files.

However, text/template/parse did not skip it. This is usually fine, as
getting the BOM prefix in the template output is harmless.

However, if the BOM byte appeared in a block overlay template, the
behavior of executing the template can change drastically. In the added
test case, the presence of the leading BOM would make the output be just
the BOM character, which many users mistook as no output.

There's really no reason to keep the BOM byte around, and it's already
skipped in Go source files, so let's do the same here.

Fixes #28482.

Change-Id: I754746fd62fb78bb2331292967f5431ccc3b8594