Update patch set 1
Patch Set 1:

I don't think that skipping the BOM in Go source files is the same as skipping it in text/template/parse.  For text/template I think BOM issues should be handled at a different level.  For Go source files there is no other level.

If we were going to do this for templates at all, which I don't think we should, the only place to do it would be ParseFiles.  We definitely should not skip a BOM in a string passed to (*Template).Parse.  That would be very confusing.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>