encoding/asn1: make GeneralizedTime and UTCTime follow DER restriction
X.690 requires, according to its sections regarding restrictions on
GeneralizedTime (11.7) and UTCTime (11.8), that time values being DER
encoded always have a 'Z' suffix. X.690 can be found at:
	https://www.itu.int/rec/T-REC-X.690-201508-I
The existing implementation of the marshaler does not enforce this, and,
in case of GeneralizedTime, may encode a time zone offset instead of the 'Z'.

Now, in makeBody and makeField, a time.Time value is converted to UTC
first, before it is used, just in case was associated with a location
previously. This already makes the implementation follow the DER
restrictions. In appendTimeCommon the code which encoded the time zone
offset is removed, as it is not needed anymore.

This patch also implicitely fixes edge cases near years 1950 and 2050 --
when the time is shifting out of the UTCTime range during conversion
from local to UTC time; test cases for this have been added.

Fixes #19890

Change-Id: I6d80fa19565b3dd53cb4c981560a35db576302fa