go/doc: inspect result types when synthesising playground examples
One of generated examples was broken:
https://golang.org/pkg/image/png/#example_Decode.
It did not have the "io" package imported.
This package was referenced in the result type of the function
used by the example.

The root cause: the "playExample" function did not inspect
the result types of declared functions.
This CL fixes the issue by inspecting those types.

Fixes #28492
Updates #9679

Change-Id: I6d8b11bad2db8ea8ba69039cfaa914093bdd5132