fmt: Clean up and slightly restructure Printf docs
The docs for format specifiers are a bit disorganized,
with explicit argument indexes separated from the rest of them,
mutually-exclusive claims (both "width" and "explicit argument
indexes" are stated to "immediately" preceed the verb), extra
qualifiers that don't seem to add information, missing qualifiers
that I think are useful, and so on.

This is an attempt to give it slightly better flow, fewer
direct contradictions, and a little more hinting for people
who *didn't* come from a C background. We shouldn't assume Go
will always be a new language that is being learned only by
people with prior experience in other languages.

Fixes #28509

Change-Id: I1941df03176f05ee73d914b32c817c2f0acaec03