Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > Whoops, forgot to include Ian. Ian, see previous comment.
> > > 
> > > I would not be okay with such a change, because it relies on excessive synchronization.
> > 
> > The synchronization is basically already there. In the common case the locking behavior wouldn't change at all. You'd change an "isSorted bool" in the existing RLock section.
> 
> Er, I meant "You'd check an..."

Going to have to disagree with that approach (but you are welcome to make the change yourself).  I don't like the url list being modified from two different places, and I don't like doing work under a lock in the fast path.   There are a ton of optimizations here that could be made, but this is a reasonable one as is.

For the record, having a dirty bit and sorting on each request would have even worse degenerate behavior if the prefix slice was modified on every request.  Does this happen in practice?  Does anyone have more than 1000 urls handled in a single ServeMux?   Who knows.  I don't think it's reasonable to prefer one approach over another without more data.  If experience is any guide, then the answer is "it doesn't matter".

Patch-set: 4
Reviewer: Gerrit User 12225 <12225@62eb7196-b449-3ce5-99f1-c037f21e1705>